furniture_backend/requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
psycopg==3.2.6
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.1
